(feat): use full package manager url in argument
