+++ web/src/routes/agent/+page.svelte
@@ -3,6 +3,7 @@
     import { error as pageError } from "@sveltejs/kit";
     import { page } from "$app/stores";
     import appInsights from "$lib/utils/appInsights";
+    import { validateURL } from "$lib/utils/validateURL";
 
     import ChatWindow from "$lib/components/chat/ChatWindow.svelte";
     import { type Agent } from "$lib/types/Agent";
@@ -30,13 +31,34 @@
             referrer = githubRef;
             referrer_kind = "github";
         } else if (npmRef) {
-            referrer = npmRef;
+            const { isValid, packageName } = validateURL(npmRef, "npm");
+            if (!isValid) {
+                loading = false;
+                errorMessage = "Invalid NPM URL";
+                appInsights.trackException({ error: new Error(errorMessage) }); // Track exception
+                throw pageError(400, errorMessage);
+            }
+            referrer = packageName;
             referrer_kind = "npm";
         } else if (pypiRef) {
-            referrer = pypiRef;
+            const { isValid, packageName } = validateURL(pypiRef, "pypi");
+            if (!isValid) {
+                loading = false;
+                errorMessage = "Invalid PyPI URL";
+                appInsights.trackException({ error: new Error(errorMessage) }); // Track exception
+                throw pageError(400, errorMessage);
+            }
+            referrer = packageName;
             referrer_kind = "pypi";
         } else if (pubRef) {
-            referrer = pubRef;
+            const { isValid, packageName } = validateURL(pubRef, "pub");
+            if (!isValid) {
+                loading = false;
+                errorMessage = "Invalid Pub URL";
+                appInsights.trackException({ error: new Error(errorMessage) }); // Track exception
+                throw pageError(400, errorMessage);
+            }
+            referrer = packageName;
             referrer_kind = "pub";
         } else {
             loading = false;
